Create Сбор данных по API Python
This Python script uses the requests library to fetch country data from the restcountries.com API in JSON format. The data includes country codes, names, capitals, regions, subregions, languages, area, and population. The script checks if the request is successful and writes the data into a CSV file using the csv module. All the information is saved in the file countries_data.csv.